code/deployment/app/requirements.txt
streamlit==1.49.1
streamlit-drawable-canvas==0.9.3
requests==2.32.5
Pillow==11.3.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
